requirements.txt
Flask==2.2.5
Flask-Session==0.4.0
SQLAlchemy==1.4.17
Werkzeug==2.2.2
cs50==9.0.0
face-recognition==1.3.0
Pillow==9.0.0